Release 2.3.3
- Removed Yatopia Leaflight since it merged to master.

- Fixed can't find log4j2.xml error.

- Fixed 281 not expected here error.

- Changed default minimum RAM to 1024K from 1536M.

Minimum RAM is minimum RAM; some say unused RAM is wasted RAM,
but that's not used memory either. It forces JVM to use 1.5GB
memory at all times under all conditions, so even if it needs just
only 300 or 500 MB of RAM, it uses 3x of that for no reason.

- Added a message before checking for updates and added a setting
to disable it.

- Maked the log4j.configurationFile JVM system property point to a relative
path to not expose logged-in user account in Timings reports or such.

- Removed "Using patched JAR file" verbose message.

- Removed -XX:+DisableExplicitGC JVM argument.

- Removed duplicate JVM arguments about -Xms and -Xmx.
The options -XX:InitialHeapSize and -XX:MaxHeapSize are an alias for
-Xms and -Xmx, removed them.

- Removed the jdk.allowAttachSelf JVM system property.

- Combined the startup script with aikar's arguments and Mojang's
recommended defaults.

- Re-added CMS and lock optimization flags to improve performance on
< Java 15.

- Make io_timeout default to 30 which is Paper default.

- Added "agressive" pre-made option set.

- Fixed OS Name not changing after upgrading/downgrading OS.

- Added Yatoclip support.

- Switched automatic Yatopia downloader to download latest stable build.

- The startup script now downloads Java automatically if it can't find one.

Signed-off-by: Mustafa ÖNCEL <support@lifemcserver.com>